requirements.txt
flask
flask_cors
whisper
python-dotenv
openai
requests
werkzeug 